CLN : literalinclude tree editor example
	new file:   docs/source/traitsui_user_manual/examples/tree_editor.py
	modified:   docs/source/traitsui_user_manual/factories_advanced_extra.rst
	deleted:    examples/tutorials/doc_examples/examples/tree_editor.py
